Non-int versions of `Point` and `Displacement` (#4824)
* Non-int versions of `Point` and `Displacement`

This will allow us to make some structs, such as `ActorPosition`, much
smaller.

* ActorPosition: Use smaller types

`Monsters`: 56K -> 46K

* player.cpp: Reduce size of `DirectionSettings`

* CrawlTable: Displacement -> DisplacementOf<int8_t>

* CrawlTable: vector<vector> -> array<vector>

Also only allocate one vector during construction instead of two.
A bit less indirection.

* Monster#enemyPosition: Point -> WorldTilePosition

sizeof(Monster): 240 -> 232

* Monster: Further optimize field layout and sizes

sizeof(Monster): 232 -> 208

`Monsters` is down to 40,000 bytes

* DMonsterStr: _mx/_my -> position

+++ Source/player.h
@@ -813,10 +813,10 @@ void PlayDungMsgs();
 
 /* data */
 
-extern int plrxoff[9];
-extern int plryoff[9];
-extern int plrxoff2[9];
-extern int plryoff2[9];
+extern int8_t plrxoff[9];
+extern int8_t plryoff[9];
+extern int8_t plrxoff2[9];
+extern int8_t plryoff2[9];
 extern int StrengthTbl[enum_size<HeroClass>::value];
 extern int MagicTbl[enum_size<HeroClass>::value];
 extern int DexterityTbl[enum_size<HeroClass>::value];
